:sparkles: - Adds unit test slug page
